requirements.txt
google-generativeai
gtts
langdetect
pygame
sounddevice
numpy
pyttsx3
SpeechRecognition
